Fixed more ranking issues in the main leaderboad + other additions
Added more code to the point system - is till not complete or working.

Ranking seems better, with new code logic. Cleaned up some code to add to the leaderboard. It now compiles the information in the text boxes and sends it to a method designed to write to the data grid.